moved assignment notification filters to new class
